Task 38 delete comment api

+++ src/main/kotlin/community/mingle/mingledemo/domain/post/facade/PostFacade.kt
@@ -6,12 +6,13 @@ import community.mingle.mingledemo.domain.post.service.PostService
 import community.mingle.mingledemo.dto.post.PostDetailDto
 import community.mingle.mingledemo.dto.post.PostDto
 import community.mingle.mingledemo.dto.post.PostPreviewDto
+import community.mingle.mingledemo.dto.post.util.PostDtoUtil.isMyPost
 import community.mingle.mingledemo.dto.post.util.PostDtoUtil.toDetailDto
 import community.mingle.mingledemo.dto.post.util.PostDtoUtil.toDto
 import community.mingle.mingledemo.dto.post.util.PostDtoUtil.toPreviewDtos
 import community.mingle.mingledemo.enums.BoardType
 import community.mingle.mingledemo.enums.CategoryType
-import community.mingle.mingledemo.exception.InvalidPostAccess
+import community.mingle.mingledemo.exception.post.InvalidPostAccessException
 import org.springframework.data.domain.PageRequest
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
@@ -80,7 +81,7 @@ class PostFacade(
                 memberId = memberId,
                 postId = postId
             )
-        ) throw InvalidPostAccess()
+        ) throw InvalidPostAccessException()
 
         val post = postService.getById(postId)
         postService.updateViewCount(postId)
@@ -102,7 +103,7 @@ class PostFacade(
                 memberId = memberId,
                 postId = postId
             )
-        ) throw InvalidPostAccess()
+        ) throw InvalidPostAccessException()
         val post = postService.update(
             postId = postId,
             title = title,
@@ -119,8 +120,17 @@ class PostFacade(
 
     @Transactional
     fun delete(
-        postId: Long
+        postId: Long,
+        memberId: Long,
     ) {
+        val post = postService.getById(postId)
+        val member = memberService.getById(memberId)
+        if (!isMyPost(
+                post = post,
+                member = member
+            )
+        ) throw InvalidPostAccessException()
+
         postService.delete(postId)
         postImageService.delete(postId)
     }
